extracted load save menu into its own class for readability and it makes more sense for it to not be in main
